requirements.txt
torch>=2.2.0
torchvision>=0.13.0
Pillow>=8.3.1